requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
feedgen==1.0.0
markdown==3.5.2
google-generativeai==0.3.2
GitPython==3.1.42
fastapi==0.110.0
uvicorn==0.27.1
apscheduler==3.10.4
httpx==0.27.0